PROMO-267: Added support for splitOnValue to allow single price configurables.
After all configurable products are created it will split on the value that is provided. After the split it will clean up configurables that only have a single simple product assigned. After that it will readd the configurables and will create a sequence of the configurables if there are more than one of the configurable available.